#96 - 테스트 클래스 이관
