--- module-api/src/test/java/dev/be/ModuleApiApplicationTests.java
@@ -1,13 +1,0 @@
-package dev.be;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ModuleApiApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
