+++ module-api/src/test/java/dev/be/moduleapi/ModuleApiApplicationTests.java
@@ -1,0 +1,13 @@
+package dev.be.moduleapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ModuleApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
