+++ module-api/src/test/java/dev/be/moduleapi/config/TestJpaConfig.java
@@ -1,0 +1,18 @@
+package dev.be.moduleapi.config;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.domain.AuditorAware;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+import java.util.Optional;
+
+@EnableJpaAuditing
+@TestConfiguration
+public class TestJpaConfig {
+
+    @Bean
+    AuditorAware<String> auditorAware() {
+        return () -> Optional.of("test");
+    }
+}
