+++ module-api/src/test/java/dev/be/moduleapi/place/controller/PlaceRecommendationApiControllerTest.java
@@ -1,0 +1,5 @@
+package dev.be.moduleapi.place.controller;
+
+class PlaceRecommendationApiControllerTest {
+
+}
