--- src/test/java/com/dateplanner/DatePlannerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.dateplanner;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DatePlannerApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
