--- src/test/java/com/dateplanner/place/controller/PlaceRecommendationApiControllerTest.java
@@ -1,7 +1,0 @@
-package com.dateplanner.place.controller;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class PlaceRecommendationApiControllerTest {
-
-}
